JE2-69 [FE] Implemented refactoring in ProductForm.
